% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_sankey.R
\name{excel_sankey}
\alias{excel_sankey}
\title{Creates Sankey Diagram to show how K groups of a measure change over N intervals}
\usage{
excel_sankey(df, with_border = TRUE)
}
\arguments{
\item{df}{KxN dataframe i.e. risk measures for K risk factors over N timepoints}

\item{with_border}{Include white vertical and horizontal borders between different stratum and alluvium}
}
\description{
Creates Sankey Diagram to show how K groups of a measure change over N intervals
}
